@@ -1,4 +1,4 @@
-# Copyright 2023 Oreum Industries
+# Copyright 2024 Oreum Industries
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -134,28 +134,32 @@ def get_log_jcd_scan(
 
 def calc_f_beta(precision: np.array, recall: np.array, beta: float = 1.0) -> np.array:
     """Set beta such that recall is beta times more important than precision"""
-    return (1 + beta**2) * (precision * recall) / ((beta**2 * precision) + recall)
+    with np.errstate(divide='ignore', invalid='ignore'):
+        fb = (1 + beta**2) * (precision * recall) / ((beta**2 * precision) + recall)
+    return np.nan_to_num(fb, nan=0, posinf=0, neginf=0)
 
 
 def calc_binary_performance_measures(y: np.array, yhat: np.array) -> pd.DataFrame:
     """Calculate tpr (recall), fpr, precision, accuracy for binary target,
-    using all samples from PPC, use vectorised calcs
+    using quantiles of all samples from PPC, use vectorised calcs
     shapes y: (nsamples,), yhat: (nsamples, nobservations)
     """
-    yhat_pct = np.percentile(yhat, np.arange(0, 101, 1), axis=0).T
+    qs = np.round(np.arange(0, 1.01, 0.01), 2)
+    yhat_q = np.quantile(yhat, qs, axis=0, method='linear').T
     y_mx = np.tile(y.reshape(-1, 1), 101)
 
     # calc tp, fp, tn, fn vectorized
-    tp = np.nansum(np.where(yhat_pct == 1, y_mx, np.nan), axis=0)
-    fp = np.nansum(np.where(yhat_pct == 1, 1 - y_mx, np.nan), axis=0)
-    tn = np.nansum(np.where(yhat_pct == 0, 1 - y_mx, np.nan), axis=0)
-    fn = np.nansum(np.where(yhat_pct == 0, y_mx, np.nan), axis=0)
+    tp = np.nansum(np.where(yhat_q == 1, y_mx, np.nan), axis=0)
+    fp = np.nansum(np.where(yhat_q == 1, 1 - y_mx, np.nan), axis=0)
+    tn = np.nansum(np.where(yhat_q == 0, 1 - y_mx, np.nan), axis=0)
+    fn = np.nansum(np.where(yhat_q == 0, y_mx, np.nan), axis=0)
 
     # calc tpr (recall), fpr, precision etc
-    accuracy = (tp + tn) / (tp + tn + fp + fn)
-    tpr = recall = tp / (tp + fn)
-    fpr = fp / (tn + fp)
-    precision = np.nan_to_num(tp / (tp + fp), nan=1)  # beware of divide by zero
+    with np.errstate(divide='ignore', invalid='ignore'):
+        accuracy = (tp + tn) / (tp + tn + fp + fn)
+        tpr = recall = tp / (tp + fn)
+        fpr = fp / (tn + fp)
+        precision = np.nan_to_num(tp / (tp + fp), nan=1)  # beware of divide by zero
 
     perf = pd.DataFrame(
         {
@@ -168,15 +172,15 @@ def calc_binary_performance_measures(y: np.array, yhat: np.array) -> pd.DataFram
             'f1': calc_f_beta(precision, recall, beta=1),
             'f2': calc_f_beta(precision, recall, beta=2),
         },
-        index=np.arange(101),
+        index=qs,
     )
-    perf.index.set_names('pct', inplace=True)
+    perf.index.set_names('q', inplace=True)
 
-    return perf
+    return perf.round(6)
 
 
 def calc_mse(y: np.ndarray, yhat: np.ndarray) -> tuple[np.ndarray, pd.Series]:
-    r""" Convenience: Calculate MSE using all samples
+    r""" Convenience: Calculate MSE using quantiles of all samples from PPC
         y shape: (nobs, )
         yhat shape: (nsamples, nobs)
 
@@ -204,23 +208,23 @@ def calc_mse(y: np.ndarray, yhat: np.ndarray) -> tuple[np.ndarray, pd.Series]:
     mse = np.mean(se, axis=0)  # 1
 
     # collapse samples to a range of summary stats then calc error
-    smry = np.arange(0, 101, 2)
-    se_pct = np.power(np.percentile(yhat, smry, axis=0) - y, 2)  # (len(smry), nobs)
-    mse_pct = np.mean(se_pct, axis=1)  # len(smry)
+    qs = np.round(np.arange(0, 1.01, 0.01), 2)
+    se_q = np.power(np.quantile(yhat, qs, axis=0) - y, 2)  # (len(smry), nobs)
+    mse_q = np.mean(se_q, axis=1)  # len(smry)
 
-    s_mse_pct = pd.Series(mse_pct, index=smry, name='mse')
-    s_mse_pct.index.rename('pct', inplace=True)
-    return mse, s_mse_pct
+    s_mse_q = pd.Series(mse_q, index=qs, name='mse')
+    s_mse_q.index.rename('q', inplace=True)
+    return mse, s_mse_q
 
 
 def calc_rmse(y: np.ndarray, yhat: np.ndarray) -> tuple[np.ndarray, pd.Series]:
-    """Convenience: Calculate RMSE using all samples
+    """Convenience: Calculate RMSE using quantiles of all samples from PPC
     shape (nsamples, nobs)
     """
-    mse, s_mse_pct = calc_mse(y, yhat)
-    s_rmse_pct = s_mse_pct.map(np.sqrt)
-    s_rmse_pct._set_name('rmse', inplace=True)
-    return np.sqrt(mse), s_rmse_pct
+    mse, s_mse_q = calc_mse(y, yhat)
+    s_rmse_q = s_mse_q.map(np.sqrt)
+    s_rmse_q._set_name('rmse', inplace=True)
+    return np.sqrt(mse), s_rmse_q
 
 
 def calc_r2(y: np.ndarray, yhat: np.ndarray) -> tuple[np.ndarray, pd.Series]:
