@@ -1,4 +1,4 @@
-# Copyright 2023 Oreum Industries
+# Copyright 2024 Oreum Industries
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -112,7 +112,7 @@ def forestplot_single(
     mdlname = kwargs.pop('mdlname', None)
     txtadd = kwargs.pop('txtadd', None)
     clr_offset = kwargs.pop('clr_offset', 0)
-    dp = kwargs.pop('dp', 1)
+    dp = kwargs.pop('dp', 2)
     plot_med = kwargs.pop('plot_med', True)
     plot_combined = kwargs.pop('plot_combined', False)
     kws = dict(
@@ -123,9 +123,9 @@ def forestplot_single(
 
     qs = np.quantile(data, q=[0.03, 0.25, 0.5, 0.75, 0.97])
     desc = (
-        f'med {qs[2]:.{dp}f}, HDI50 ['
+        f'\nOverall: Median {qs[2]:.{dp}f}, $HDI_{{50}}$ = ['
         + ', '.join([f'{qs[v]:.{dp}f}' for v in [1, 3]])
-        + '], HDI94 ['
+        + '], $HDI_{94}$ = ['
         + ', '.join([f'{qs[v]:.{dp}f}' for v in [0, 4]])
         + ']'
     )
@@ -136,7 +136,7 @@ def forestplot_single(
     if plot_med:
         _ = ax0.axvline(qs[2], color='#ADD8E6', ls='--', lw=3, zorder=-1)
     _ = f.suptitle(
-        ' - '.join(filter(None, ['Forestplot levels', mdlname, group, txtadd, desc]))
+        ' - '.join(filter(None, ['Forestplot levels', mdlname, group, txtadd])) + desc
     )
     _ = f.tight_layout()
     return f
@@ -253,11 +253,12 @@ def plot_ppc(
     ecdf: bool = True,
     data_pairs: dict = None,
     flatten: list = None,
+    observed_rug: bool = True,
     **kwargs,
 ) -> figure.Figure:
     """Plot In- or Out-of-Sample Prior or Posterior predictive ECDF, does not
     require log-likelihood.
-    NOTE: data_pairs {key (in observed): value (in {group}_predictive)}
+    NOTE: data_pairs {key (in observed_data): value (in {group}_predictive)}
     """
     txtadd = kwargs.pop('txtadd', None)
     kind = 'cumulative' if ecdf else 'kde'
@@ -276,6 +277,8 @@ def plot_ppc(
         ax=axs,
         data_pairs=data_pairs,
         flatten=flatten,
+        observed_rug=observed_rug,
+        random_seed=42,
         **kwargs,
     )
     t = f'{"In" if insamp else "Out-of"}-sample {group.title()} Predictive {kindnm}'
@@ -373,7 +376,7 @@ def plot_lkjcc_corr(mdl: BasePYMCModel, **kwargs) -> figure.Figure:
     return facetplot_krushke(
         mdl=mdl,
         txtadd='lkjcc_corr, diagonals only',
-        rvs=mdl.rvs_corr,
+        rvs=['lkjcc_corr'],
         coords=coords,
         m=2,
         rvs_hack=0,
